split struclog.h into log.h and log_sink.h
