+++ modules/telemetry/telemetry/CMakeLists.txt
@@ -11,8 +11,8 @@ declare_module(
 find_package(fmt REQUIRED)
 
 # library sources
-set(SOURCES src/metric_sink.cpp src/metric_record.cpp src/struclog.cpp src/log_sinks/absl_sink.cpp src/metric_sinks/terminal_sink.cpp 
-            include/hephaestus/telemetry/metric_sink.h include/hephaestus/telemetry/metric_record.h include/hephaestus/telemetry/struclog.h include/hephaestus/telemetry/log_sinks/absl_sink.h include/hephaestus/telemetry/metric_sinks/terminal_sink.h
+set(SOURCES src/metric_sink.cpp src/metric_record.cpp src/log_sink.cpp  src/log.cpp src/log_sinks/absl_sink.cpp src/metric_sinks/terminal_sink.cpp 
+            include/hephaestus/telemetry/metric_sink.h include/hephaestus/telemetry/metric_record.h include/hephaestus/telemetry/log_sink.h include/hephaestus/telemetry/log.h include/hephaestus/telemetry/log_sinks/absl_sink.h include/hephaestus/telemetry/metric_sinks/terminal_sink.h
 )
 
 # library target
