+++ modules/telemetry/telemetry/tests/CMakeLists.txt
@@ -10,8 +10,8 @@ define_module_test(
         PUBLIC_LINK_LIBS "")
 
 define_module_test(
-        NAME struclog_tests
-        SOURCES struclog_tests.cpp
+        NAME log_tests
+        SOURCES log_tests.cpp
         PUBLIC_INCLUDE_PATHS
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../include>
         PUBLIC_LINK_LIBS "")
